ATO-827: Creates a oidc provider
Creates an oidc provider using the parameters from the node-oidc-provider
package. This tries to mirror our own oidc configuration[1] as closely
as possible. This probably still needs a bit more work but is a start
the provider uses the loadExsistingGrant function which is outlined
in the documentation[1] and is adapted from some example code[2]

[1] - https://github.com/panva/node-oidc-provider/tree/main/docs#interactions
[2] - https://github.com/panva/node-oidc-provider/blob/main/recipes/skip_consent.md