ATO-827: Adds private keys to provider config
Adds the private keys to the provider config. This removes a warning
about using the default keys and also enables both signing algorithms
we have specified in the config.

!NOTE - If there is not a valid EC key in the config, the signing
algorithm ES256 is disabled, despite it being specified in the
provider configuration. I could not find documentation for this
in the package docs, so have added it as a note here.