ATO-827: Adds scope validator
Adds a scope validator which attempts to mimic our own oidc
setup. Small caveat that by default this package seems to ignore
scopes which are not known to the provider, which matches what is
outlined in the specification here[1]. However this behaviour is not
what we do, so this uses the url query param instead of the scope
value that will be passed to the validator as the package drops
any scopes which are not known to the provider from this argument.

[1]- https://openid.net/specs/openid-connect-core-1_0.html#:~:text=Authorization%20Code%20Flow%3A-,scope,-REQUIRED.%20OpenID%20Connect