ATO-933: Adds constants for token creation
This adds the constant values for needed for token creation,
like the trustmark uri, expiry values and signing keys.

!NOTE- This commit includes two private keys. This is fine, as these keys
are just used to sign the access tokens issued by the simulator's /token
endpoint. The simulator is a test product, and therefore these keys do
not need to be a secret.