ATO-827: Adds dedupeQueryParams middleware
Our API Gateway will ignore duplicate query parameters and just use
the last instance of the key-value pair[1]. Express will not do this and
parses duplicate query params as an array of values. Express uses the qs
library for parsing query strings which also has support for parsing
objects from query parameters[2]. We need to disregard objects in the
query params and emulate the behaviour of API gateway, so this middleware
achieves that.

[1]- https://aws.amazon.com/blogs/compute/support-for-multi-value-parameters-in-amazon-api-gateway/#:~:text=You%20can%20see%20the%20original%20behavior%20in%20the%C2%A0queryStringParameters%C2%A0parameter%20in%20the%20above%20input%2C%20where%20only%20the%20%E2%80%9Cfish%E2%80%9D%20value%20is%20retained.
[2] - https://www.npmjs.com/package/qs