BAU: Fix VTR logic for P0 cases
The previous logic was incorrrect for parsing a VTR containing P0
and a low level of credential trust. The previous logic:

```
(parsedVtr.levelOfConfidence !== null ||
      parsedVtr.levelOfConfidence !== "P0") &&
    parsedVtr.credentialTrust !== "Cl.Cm"
```

would return true in the case P0 and Cl and therefore this function would
would throw when supplied with a P0 and low level of credentential
trust.

I have ameded the logic to be more readable and explicit for what
we are looking for, namely if an identity vector has been requested
and the credential trust is not medium. I have also added a test
which asserts what the behaviour should be, namely we accept any
level of credential trust when using P0