ATO-1441: Add INTERACTIVE_MODE flag
RPs have indicated that they would like the simulator to be
able to return multiple response configurations for a given
request, in parallel, for example using browser driven e2e
tests. To accommodate this we will introduce a new mode we
will call 'interactive  mode' which will render a form to
collect a given request's response configuration.