BAU: Refactor test setup
There was previously a lot of duplicated test code which setup env
vars and keys, as well as a repeated constant for nonce and state.
This refactors the test code to reduce a fair amount of duplication
and saves ~100 lines of code. This diff for this is a bit much,
but there are no functional changes in this commit, just
re-organising the code