ATO-1443: Add encoded request params and state in form
We need both the state and the auth request params to be able to
store against the embded auth code upon form submission. This
stores them both as hidden form inputs to be able to persist them
when rendering and submitting the form.

The Auth request params are base64Url encoded as this prevents '/'
characters from escaping the value attribute and messing up the
rendering of the form.