ATO-1397: Add hi-base32 depedency
For running the acceptance tests against integration, we're gonna
switch to using an auth app user. To allow this we need to process
a base32 string. Node does not natively support this, so we're adding
this library to do this - because handrolling this would be more
effort than it's worth