ATO-1397: Allow null as a phone number
It is possible that we sometimes don't return a phone number,
like in the case of an auth app user. The current code in the userinfo
controller behaves correctly by checking using ?? to coalecse to
underfined in null-ish cases.

The problem is currently, the config endpoint expects it to be
either not present OR a string. There is no allowance for null,
meaning you can't turn "off" the phone number using the config
endpoint. Especially considering that the config class defaults
to a value, this makes it quite hard to reproduce.

This fixes that by allowing null