ATO-1397: Rejig micro-rp to present like stub
We're gonna re-jig the tests to allow running against the micro-rp
as a stub clone instead. This will require us to format information
from the micro-rp in a similar fashion to the RP stub.

I've stolen like 90% of this code from the stub anyway, mainly the
the HTML.

I've also moved some stuff, such as keys, into environment variables.