ATO-1397: Updates actions for runinng against micro rp
Updates the github actions to use the docker compose
file to spin up the micro rp. This was because we were
seeing issues with selenium accessing the micro rp when
it was spun up seperately to the service containers in the
action.

Also adds new environment variables etc required for running the
tests