BAU: Add code to validate an Argon2 hash
We need to be able to compare a Argon2 hashed secret to the raw
secret provided in the token request as a means of client
authentication. To do this we need to be able to parse a hashed
secret back into it's params and use these params to hash the
raw secret provided.

Once we have hashed the raw secret using the same params, we need
to do a constant time comparison of each byte to ensure we match
the values.

Most of this code was taken directly from our actual implementation[1]
with some small modifications

[1] - https://github.com/govuk-one-login/authentication-api/blob/08fd7af584874bcb0fd753330e2db6a67fd8f093/orchestration-shared/src/main/java/uk/gov/di/orchestration/shared/helpers/Argon2MatcherHelper.java#L9